Dual-license mempool under both the GNU AGPLv3 and GNU GPLv3 licenses
A few weeks ago in #388, we changed the license for the mempool project
from MIT+CC to AGPLv3. This was generally met with very positive
feedback from the Bitcoin community. However, a few very large
enterprise organizations that love the mempool project are now no longer
able to use our code, since apparently some enterprise organizations
have an internal policy of not using AGPL licensed software.

Since the AGPLv3 license is compatible with the GPLv3 license in various
ways, adding the GPLv3 as an alternative second license for the project
seems like a reasonable way to retain most of the copyleft properties of
our current AGPLv3 license, while also keeping the community happy and
allowing them GPLv3 as another choice if they are prohibited from using
the AGPL internally.

Therefore, I propose to add GPLv3 as an alternative license for the
mempool open source project in this PR. If you are okay with this,
please ACK with "I hereby re-license my contributions to the mempool
open source project under both the GNU AGPLv3 and GPLv3 licenses"

- [ ] @softsimon
- [ ] @wiz
- [ ] @bguillaumat
- [ ] @TechMiX
- [ ] @rbrooklyn
- [ ] @junderw
- [ ] @andrerfneves
- [ ] @andrewtoth
- [ ] @6102bitcoin
- [ ] @Czino
- [ ] @devinbileck
- [ ] @knorrium
- [ ] @jambolo
- [ ] @lucasmoten
- [ ] @fiatjaf
- [ ] @pox
- [ ] @RandyMcMillan
- [ ] @timlucmiptev